refactor: move relation methods
